# frontend/openapi.json
{
  "components": {
    "schemas": {
      "AblationRow": {
        "properties": {
          "ablation_name": {
            "title": "Ablation Name",
            "type": "string"
          },
          "citation_validity_deterministic": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Citation Validity Deterministic"
          },
          "citation_validity_semantic": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Citation Validity Semantic"
          },
          "configuration": {
            "title": "Configuration",
            "type": "string"
          },
          "note_quality_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note Quality Score"
          },
          "triage_accuracy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Triage Accuracy"
          }
        },
        "required": [
          "ablation_name",
          "configuration",
          "citation_validity_deterministic",
          "citation_validity_semantic",
          "triage_accuracy",
          "note_quality_score"
        ],
        "title": "AblationRow",
        "type": "object"
      },
      "AgentMetrics": {
        "description": "Null where nothing has been measured yet - never a zero standing in for missing.",
        "properties": {
          "avg_latency_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Latency Seconds"
          },
          "avg_tokens": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Tokens"
          },
          "avg_tool_calls": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Tool Calls"
          },
          "citation_validity_deterministic": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Citation Validity Deterministic"
          },
          "citation_validity_semantic": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Citation Validity Semantic"
          },
          "notes": {
            "title": "Notes",
            "type": "integer"
          },
          "notes_failed_after_retries": {
            "title": "Notes Failed After Retries",
            "type": "integer"
          },
          "notes_regenerated": {
            "title": "Notes Regenerated",
            "type": "integer"
          },
          "triage_accuracy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Triage Accuracy"
          }
        },
        "required": [
          "notes",
          "citation_validity_deterministic",
          "citation_validity_semantic",
          "triage_accuracy",
          "avg_tool_calls",
          "avg_tokens",
          "avg_latency_seconds",
          "notes_regenerated",
          "notes_failed_after_retries"
        ],
        "title": "AgentMetrics",
        "type": "object"
      },
...
      "DetectorMetrics": {
        "properties": {
          "anomaly_type": {
            "title": "Anomaly Type",
            "type": "string"
          },
          "detector": {
            "title": "Detector",
            "type": "string"
          },
          "f1": {
            "title": "F1",
            "type": "number"
          },
          "fn": {
            "title": "Fn",
            "type": "integer"
          },
          "fp": {
            "title": "Fp",
            "type": "integer"
          },
          "granularity": {
            "enum": [
              "case",
              "row"
            ],
            "title": "Granularity",
            "type": "string"
          },
          "pr_auc": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pr Auc"
          },
          "precision": {
            "title": "Precision",
            "type": "number"
          },
          "recall": {
            "title": "Recall",
            "type": "number"
          },
          "tp": {
            "title": "Tp",
            "type": "integer"
          }
        },
        "required": [
          "detector",
          "anomaly_type",
          "granularity",
          "precision",
          "recall",
          "f1",
          "pr_auc",
          "tp",
          "fp",
          "fn"
        ],
        "title": "DetectorMetrics",
        "type": "object"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "EvaluationResponse": {
        "properties": {
          "ablations": {
            "items": {
              "$ref": "#/components/schemas/AblationRow"
            },
            "title": "Ablations",
            "type": "array"
          },
          "agent_metrics": {
            "$ref": "#/components/schemas/AgentMetrics"
          },
          "baseline_metrics": {
            "items": {
              "$ref": "#/components/schemas/DetectorMetrics"
            },
            "title": "Baseline Metrics",
            "type": "array"
          },
          "dataset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset"
          },
          "detector_metrics": {
            "items": {
              "$ref": "#/components/schemas/DetectorMetrics"
            },
            "title": "Detector Metrics",
            "type": "array"
          },
          "generated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated At"
          },
          "injected_anomaly_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Injected Anomaly Count"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "seed": {
            "title": "Seed",
            "type": "integer"
          }
        },
        "required": [
          "run_id",
          "seed",
          "dataset",
          "injected_anomaly_count",
          "detector_metrics",
          "baseline_metrics",
          "agent_metrics",
          "ablations",
          "generated_at"
        ],
        "title": "EvaluationResponse",
        "type": "object"
      },
...
      "RunSummary": {
        "properties": {
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "kind": {
            "enum": [
              "ingest",
              "detect",
              "investigate",
              "evaluate"
            ],
            "title": "Kind",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "enum": [
              "running",
              "completed",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "additionalProperties": true,
            "title": "Summary",
            "type": "object"
          }
        },
        "required": [
          "run_id",
          "kind",
          "status",
          "seed",
          "started_at",
          "finished_at",
          "summary"
        ],
        "title": "RunSummary",
        "type": "object"
      },
...
      "TransactionRow": {
        "properties": {
          "amount": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount",
            "type": "string"
          },
          "cited": {
            "title": "Cited",
            "type": "boolean"
          },
          "in_case": {
            "title": "In Case",
            "type": "boolean"
          },
          "invoice_no": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invoice No"
          },
          "item_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Category"
          },
          "item_desc": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Desc"
          },
          "officer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Officer Id"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "row_id": {
            "title": "Row Id",
            "type": "integer"
          },
          "txn_date": {
            "format": "date",
            "title": "Txn Date",
            "type": "string"
          },
          "unit_price": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Price"
          },
          "vendor_key": {
            "title": "Vendor Key",
            "type": "string"
          },
          "vendor_name": {
            "title": "Vendor Name",
            "type": "string"
          }
        },
        "required": [
          "row_id",
          "vendor_name",
          "vendor_key",
          "invoice_no",
          "amount",
          "txn_date",
          "officer_id",
          "item_desc",
          "item_category",
          "quantity",
          "unit_price",
          "in_case",
          "cited"
        ],
        "title": "TransactionRow",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Procurement cases, verified audit notes, evidence and review. Reads batch results; the only writes are a reviewer's status change and the demo.",
    "title": "SpendGuard API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/evaluation": {
      "get": {
        "description": "Detector metrics against the baseline, agent metrics and ablations for one seed (FR-5.8).",
        "operationId": "evaluation_api_v1_evaluation_get",
        "parameters": [
          {
            "in": "query",
            "name": "seed",
            "required": false,
            "schema": {
              "default": 42,
              "title": "Seed",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Evaluation",
        "tags": [
          "overview"
        ]
      }
    },
...
    "/api/v1/runs": {
      "get": {
        "description": "Run history, newest first, so the dashboard can say which run it shows.",
        "operationId": "runs_api_v1_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunSummary"
                  },
                  "title": "Response Runs Api V1 Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Runs",
        "tags": [
          "overview"
        ]
      }
    },
    "/api/v1/transactions/{row_id}": {
      "get": {
        "description": "One row, for inspecting a citation. Read from the audit view: never the answer key.",
        "operationId": "transaction_api_v1_transactions__row_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "row_id",
            "required": true,
            "schema": {
              "title": "Row Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionRow"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Unknown row"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Transaction",
        "tags": [
          "overview"
        ]
      }
    }
  }
}